[pentest] Debug tests from chip tests
Changed all *_init expected responses in the testOS tests to expect clock_jitter to be disabled. Removed fields for fi_lc_ctrl, fi_ibex, and fi_crypto tests since they are random or different for each device. In addition, the tests caught a bug in fi_ibex which is fixed.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>