[pentest] Add combinations of tests
Add a draft of a bunch of branch tests.
Add an SCA test for Ibex which combines operations such as XOR, ADD,
shift, mul, and the loading and storing of data to SRAM.
Add testing for the new combinational fi ibex test.
Add more functions providing the state of the chip.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>