[host/pentest] Add Python test framework
Add Python communication/host scripts for the penetration test OS from
//sw/device/tests/penetrationtests.
In addition, test their correctness using py_tests where the latest
binaries from the test OS  are built and tested against their
communication scripts.

Since the tests contain a Python test framework, a custom Bazel
definition is needed to pull the correct silicon/fpga target. This
Bazel definition only affects the pentest tests.

These tests should now also be pulled in CI.

In order to work with the FPGA and the Python framework, chipwhisperer needs to be pulled in as a dependency.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>